requirements.txt
streamlit==0.75.0
scikit-learn==1.0.1
Pillow~=9.0.0